Feature/backend

+++ Backend/doodle/src/main/java/com/cyworld/minihompy/entity/Post.java
@@ -11,7 +11,8 @@ import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
-@Getter @Setter
+@Getter
+@Setter
 @NoArgsConstructor
 @Table(name = "Posts")
 @EntityListeners(AuditingEntityListener.class)
@@ -38,7 +39,7 @@ public class Post {
     private LocalDateTime updatedAt;
 
     @Column(nullable = false)
-    private Integer viewCount = 0;
+    private Long viewCount = 0L;
 
     @Column(nullable = false)
     private Boolean isDeleted = false;
